backend/report_service/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
strawberry-graphql[fastapi]==0.245.0
httpx==0.27.2
# Data Validation
pydantic==2.9.2
pydantic-settings==2.6.0
# Environment Variables
python-dotenv==1.0.1
# Utils
python-multipart==0.0.12
python-dateutil==2.9.0
# GraphQL core
graphql-core==3.2.3
typing-extensions==4.12.2